feat(container): add function to close SNOW session
ref: MANAGER-15587
